requirements.txt
# ===============================
# CORE FLASK
# ===============================
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0

# ===============================
# SECURITY
# ===============================
bcrypt==4.0.1
Werkzeug==2.3.7

# ===============================
# DATABASE
# ===============================
psycopg2-binary==2.9.9

# ===============================
# SERVER
# ===============================
gunicorn==21.2.0
python-dotenv==1.0.0

# ===============================
# REPORTING / EXPORTS
# ===============================
reportlab==4.0.9
openpyxl==3.1.5
matplotlib==3.9.2
pytz==2024.2

# ===============================
# AWS / FINOPS
# ===============================
boto3==1.28.62

# ===============================
# DATA (FUTURO / BI)
# ===============================
pandas==2.2.3
numpy==2.1.2

# ===============================
# TEMPLATES
# ===============================
Jinja2==3.1.2
